Use filters as array in dxTagBox
