+++ testing/tests/DevExpress.ui.widgets.editors/tagBox.tests.js
@@ -4357,7 +4357,7 @@ QUnit.test("Select All should use cache", function(assert) {
         keyGetterCounter++;
         return this._id;
     };
-    for(var i = 1; i <= 10; i++) {
+    for(var i = 1; i <= 100; i++) {
         var item = { _id: i, text: "item " + i };
         Object.defineProperty(item, "id", {
             get: getter,
@@ -4389,7 +4389,7 @@ QUnit.test("Select All should use cache", function(assert) {
     $(".dx-list-select-all-checkbox").trigger("dxclick");
 
     //assert
-    assert.equal(keyGetterCounter, 144, "key getter call count");
+    assert.equal(keyGetterCounter, 1404, "key getter call count");
     assert.equal(isValueEqualsSpy.callCount, 0, "_isValueEquals is not called");
 });
